app/src/main/java/com/picpay/desafio/android/data/local/UserDao.kt
package com.picpay.desafio.android.data.local

import androidx.room.Dao
import androidx.room.Delete
import androidx.room.Insert
import androidx.room.OnConflictStrategy
import androidx.room.Query
import com.picpay.desafio.android.data.models.UserEntity
import kotlinx.coroutines.flow.Flow

@Dao
interface UserDao {

    @Insert(onConflict = OnConflictStrategy.REPLACE)
    suspend fun insert(users: List<UserEntity>)

    @Delete
    suspend fun delete(user: UserEntity)

    @Query("SELECT * from users")
    fun getAllUsers(): Flow<List<UserEntity>>

}